Add configflow step to set name of weather entity.
See #95. This commit adds a new step in the config flow to allow the
user to accept or override the name of the weather entity,
independently of whether they untick the box for forecast sensors.
The forecast location name is used as the default value for
"weather_name".

Previously, "forecast_basename" config variable is used to set
the name of the weather entity as well as the basename of additional
forecast sensors. If the user asks for no forecast sensors the user
would not be given the option to set a name.