Add integration configuration and device registry entries
Adds the ability to reconfigure the integration via the
integrations page and creates a device for each configured
base name. Each device will contain the sensors that use that
base name. Most sensors are set to diagnostic, so they won't appear
on auto-generated dashboards by default, but they can be added
by using the "Add to dashboard" option.

Signed-off-by: Avi Miller <me@dje.li>